+++ src/css_names.js
@@ -1,0 +1,6 @@
+export const cssNames = (...args) => {
+    const cNames = [...args]
+
+    cNames.toString = () => cNames.join(' ')
+    return cNames
+}